# Add NEON assembly for rejection sampling