… mlkem/rej_uniform.c