@hanno-becker 2024-06-04
> I should re-arrange the instructions, but I think when we have SLOTHY figures our the best instruction sequences, so I left it as future work.

Yes, please write the code as clean as possible for now -- micro-optimization will be left to SLOTHY. (And if you have time, please feel free to try that out yourself)