@cothan 2024-06-06
Okay, we will need to integrate arm runner to run this code in CI.