@hanno-becker 2024-06-06
@cothan I see CI for https://github.com/pq-code-package/mlkem-c-aarch64/pull/47/commits/9ee3e95a2486218faffb882940094f7b4e082033 failing because of linting, not because of an architecture problem.