@cothan 2024-06-10
Okay, after I removed the `simd_memcpy` and replace with `memcpy`. It works on Mac. 
I can't wrap my head why the `arm64` version land on the `#ifdef __APPLE_` branch and take on `_memcpy` while on Linux. 
@.@ I need to investigate why the compiler on `arm64` land on `__APPLE__` branch.