@cothan 2024-06-10
Okay, I know why. The assembly extension must be capital. `.S` instead of `.s`. :(( https://stackoverflow.com/a/30327226