@cothan 2024-06-28
My M1 benchmark:

**Before:**

| Parameter           | MLKEM512 | MLKEM768 | MLKEM1024 |
|---------------------|----------|----------|-----------|
| keypair cycles      | 60678    | 103056   | 161138    |
| encaps cycles       | 75188    | 122977   | 184479    |
| decaps cycles       | 93538    | 146287   | 214105    |

**After:**

| Parameter           | MLKEM512 | MLKEM768 | MLKEM1024 |
|---------------------|----------|----------|-----------|
| keypair cycles      | 59284    | 99135    | 153797    |
| encaps cycles       | 73790    | 118616   | 176734    |
| decaps cycles       | 91968    | 141692   | 206045    |