@cothan 2024-06-28
The number is much bigger (probably same ratio) compare on RPi 4. 

This branch: 
https://github.com/pq-code-package/mlkem-c-aarch64/actions/runs/9718493064?pr=47

Main branch: 
https://github.com/pq-code-package/mlkem-c-aarch64/actions/runs/9714693885