@hanno-becker 2024-07-30
Apologies @cothan, I've just discussed this with @mkannwischer again. We believe that this work is low-priority and would rather slow us down on the path towards an initial release, seeing that we strive for functional verification of all assembly. We therefore feel that the PR should be put on hold until after the first release.

I'm sorry @cothan that we did not sort this out earlier.