@hanno-becker 2024-09-27
@cothan I feel that now would be a good time to revive and update this pull request, rejection sampling being the last performance-relevant function that is not yet written in assembly. I don't know if/when we will manage to verify it, but if not, we may hide it behind an appropriate flag if needed.

Do you have time to work on this?