@cothan 2024-09-27
@hanno-becker , yup, I have time to work on this PR, I would like to use this PR to try out SLOTHY as well.