@hanno-becker 2024-09-27
@cothan Nice! Let's start with the clean version in this PR, and do an optimized one as a follow-up. You will need to add quite a few instructions to SLOTHY.